Rebased and addressed follow up comments from PR 2301
